Make `getinlinecache` less extreme
This patch refactors the way getinlinecache works.  We keep a list of
every location that refers to `getinlinecache`.  When the generated code
gets hit, we execute a stub if the global constant state is good. The
stub generates more optimized code that we jump to on subsequent calls

If the global constant state is ever changed, we walk the list of
getinlinecache references and patch them to use the slow path until the
global constant state is stable, in which case the stub is executed
again and newer optimized code generated