Implement `checkkeyword` test suite
Test suite for the `checkkeyword` instruction.

Currently, `checkkeyword` can't be tested, because it seems to require a method to be defined in the same scope; since `definemethod` is not currently implemented, the JIT exits before.

I have a `checkkeyword` implementation that could be correct, but unfortunately it requires the conditions above, so I'll put it on hold.