Use `Fiddle::Handle::DEFAULT` to find mjit symbols
It seems like Internals.symbol_address is not reliable, and I'm not sure
why.  We should use `Fiddle::Handle::DEFAULT` where we can (since it is
always correct), but also figure out why the other API is wrong.  I
assume it's something to do with ASLR, but I'm not sure.