Refactor frame pushing and stack manipulation
The callee's EP / SP is just an extension of the caller's SP.  Rather
than do a bunch of reads / register shuffling, we can pass the temporary
stack to the frame pushing code.  When we push a frame, just push the
"special values" on the stack like we'd normally push on the stack, then
do any assignments we need.  This should make the code a little more
clear and also reduce the machine code we emit.