Don't flush the SP if the target ISEQ isn't compiled yet
The deferred_call function will try to flush the SP before calling out
to the `rb_funcall` function.  In this case we have *just* pushed a
frame at runtime, and the compile time temporary stack isn't in sync
with the runtime frame push.  Since we just pushed a frame, the previous
frame's SP, and the newly pushed frame's SP should all be up to date, so
this commit adds a fake "no flush" temp stack.

We should probably come up with a different scheme, but this seemed like
the easiest fix