Fix the swap instruction
The swap instruction wasn't actually writing anything to the stack, so
this fixed it.  The other problem is that it would pop an item, then
push the item back on the stack in the reverse order.  This doesn't seem
like a problem (because that's what `swap` is supposed to do), but each
item in the temporary stack keeps track of where it needs to be read in
relation to the base pointer.

Since we just push the same item reference back on the stack, it means
that subsequent instructions that read from the stack would see the
wrong offset in relation to the base pointer.